Get interpolation values for catchment

Get interpolation values for catchment

Usage

get_interpolation(sp_df, st, grd, i, idw_exponent)

Arguments

sp_df

SpatialPointsDataFrame with one column for one day and rows as stations.

st

SpatialPointsDataFrame with virtual stations from a grid. Should contain dataframe in data for station elevation from DEM

grd

Grid for interpolation.

i

integer representing day number.

idw_exponent

numeric value for exponent parameter to be used in interpolation.

Value

SpatialPointsDataFrame with virtual stations and interpolated data in data dataframe

Examples

if (FALSE) { # \dontrun{
get_interpolation(sp_df, meteo_pts, 2, 2)
} # }