Plotting calibration data

Plotting calibration data

Usage

plot_cal_data(df, stations, variables = levels(as.factor(df$Variables)))

Arguments

df

dataframe with formatted data (Station, DATE, Variables and Values columns are needed)

stations

character vector listing stations, which should be selected for figure.

variables

optional parameter of character vector, which parameters should be in figure.

Value

plotly object of interactive figure.

Examples

temp_path <- system.file("extdata", "calibration_data.xlsx", package = "SWATprepR")
cal_data <- load_template(temp_path)
#> [1] "Loading data from template."
#> [1] "Loading of data is finished."
plot_cal_data(cal_data$data, stations = c("1", "2", "3","10"), variables =
c("PT", "Q"))