Plotting stations with few data points

Plotting stations with few data points

Usage

plot_many(df, drop_st = c())

Arguments

df

dataframe with formatted data (Station, DATE, Variables and Values columns are needed)

drop_st

optional parameter of character vector, which stations should be excluded for figure. Stations with a lot of data should be plotted with plot_one function.

Value

plotly object of interactive figure

Examples

temp_path <- system.file("templates", "calibration_data.xlsx", package = "SWATprepR")
cal_data <- load_template(temp_path)
#> [1] "Loading data from template."
#> [1] "Loading of data is finished."
plot_many(cal_data$data, drop_st = c("4"))