Plotting figure to comparing two datasets with weather data

Plotting figure to comparing two datasets with weather data

Usage

plot_weather_fig_compare(
  meteo_lst1,
  meteo_lst2,
  par,
  period = "day",
  fn_summarize = "mean",
  name_set1 = "raw",
  name_set2 = "clean"
)

Arguments

meteo_lst1

first nested list of lists with dataframes. Nested structure meteo_lst -> data -> Station ID -> Parameter -> Dataframe (DATE, PARAMETER).

meteo_lst2

second nested list of lists with dataframes. Nested structure meteo_lst -> data -> Station ID -> Parameter -> Dataframe (DATE, PARAMETER).

par

character marking weather variable to extract (i.e. "PCP", "SLR", etc).

period

character describing, which time interval to display (default is "day", other examples are "week", "month", etc).

fn_summarize

function to recalculate to time interval (default is "mean", other examples are "median", "sum", etc).

name_set1

character to name first dataset.

name_set2

character to name second dataset.

Value

plotly figure object with displayed weather data for two datasets.

Examples

##plot_weather_fig_compare(meteo_lst1, meteo_lst2, "PCP", "month", "mean", "raw", "clean")